+++ EnttPong/src/components/PositionComponent.hpp
@@ -23,7 +23,7 @@ namespace ep
 		/// \param initialX The starting x-pos on screen.
 		/// \param initialY The starting y-pos on screen.
 		///
-		PositionComponent(int initialX, int initialY);
+		PositionComponent(float initialX, float initialY);
 
 		///
 		/// Default destructor.
@@ -39,12 +39,12 @@ namespace ep
 		///
 		/// x pos on screen.
 		///
-		int m_x;
+		float m_x;
 		
 		///
 		/// y pos on screen.
 		///
-		int m_y;
+		float m_y;
 	};
 }
